Video Backgrounds now save shortcuts on EditBackground.
Added a boolean saveFile to "saveToBackgroundIDFile" ipc call which saves the file directly if saveFile = true, or saves a shortcut to the file if saveFile = false.

For now this is defaulted to True for images, and false for videos. (Saves images, saves shortcuts for videos.) Will add a way to toggle this later.

Also updated (for now) to save your current background to the newly created background (thus when EditBackground closes it opens BackgroundSelect with the new background selected).
Will likely add a toggle for this as well.