Fix unicode (Chinese Characters) in shortcut names.
The "windows-shortcuts" works great for creating shortcuts, however it has a major flaw which seems to that it, and powershell both seem to have major issues with Unicode characters. Causing constant problems with error 87 the parameter is incorrect. And trying to access files like /??/??.lnk instead of the Unicode characters.

Added a new rust executable (and source code) for a small helper executable which creates a shortcut.

The helper executable should work fine for non-rust installed computers. And since I am including the executable you do not need to compile it yourself unless you want to edit it.

Tested on my non-dev (extremely old) laptop.