nuitka wrapping file_to_image
Slightly optimizes the launch time and file size of the file_to_image.exe.
In the testing I did with this new builder, the compiled script is ~3mb smaller, and is about 30-50% faster than the pyinstaller compiled script.

Tried for a few hours to re-write this python script in a more efficient language but could not get most working and if it worked it would only work for exes, and would return an upscaled low res image not a high res image obtained from the exe then scaled, like how my python script does.

Still not happy with the performance of this. For now just decided to set it aside and revisit it later as the main program still needs a lot of work, and this slight optimization isn't worth the time (right now).