Adding a background generates a unique ID.
When creating a new background from scratch it calls EditBackground with "" for ID. When saving and this ID is set to "" it attempts to create an ID based on the name set, if no name set it gives it a number ID like "000001", "000002" etc.
If the base name already exists gives it an ID like name_1, name_2 etc.

Basically just ensuring that whenver you add a background it gets a unique ID, which it sets once, and will never update again. The name itself can change, but the ID is set once creating. So if you create a background with the name "Really cool background" and then later edit it (when its implemented) to be "Really bad background" the ID itself which is the file it is stored in is still "Really cool background". Keep in mind this is only for browsing throught the backgrounds/ files yourself. The name which displays in BackgroundSelect will show the up to date "Really bad background" name, so you can still open the file via contextMenu there.

This whole system is set up to allow users to likely find their custom backgrounds without having the program open so long as they properly name their backgrounds on creation.