More clarification on SaveToBackgroundIDFile() ipc calls.
It was working as intended, but while looking at the code it confused me as to how it was working as intended for selecting shortcut files as the background instead of the raw files. So updated the logic to specifically assign shortcuts to "true" for saving to background file. Also added logger statements to SaveToBackgroundIDFile to give more information to track something possibly misbehaving.