Resolve shortcuts on dropping. (FileSelect already does this).
Realized that in all situations which use the FolderIcon button for Selecting a file from a folder, it resolves shortcuts. This resolving is way better in general as it skips a lot of bugs that can come from saving the .lnk instead of referencing the file itself. For instance if you drop a lnk to an executable in EditBackground, you could save it and it wouldn't check the actual filetype of the target file, thus you could save a background with an executable (obviously it would not run and would fail loading on checks in Background.tsx) but it would still allow you to save it.

We can fully avoid problems like this by just referencing the target of .lnks instead. For Images/Programs in EditIcon it is better to reference the target files directly so that if the .lnk ever breaks or is deleted the reference is still good. (Slightly better than .lnks as it can still break, but copying the .lnk is twice as bad as it has 2 options to fail, target file moves or .lnk moves.)

Then the one time we want .lnks (saving video backgrounds with shorcut setting), we have our own method to create a new .lnk so no need to select an already made one.