Removed Background Path from settings. (and drag/drop).
Removed the background Path from Settings window. This was mostly a remnant from the old way of setting the background prior to BackgroundSelect. This also works way different than a user would expect (as it expects an ID, which the user cannot directly see in the program) so no reason to keep this setting. Also removed the drag/drop logic as Settings no longer needs it (at least for now).