Prevent duplicate shortcuts to same files.
Realized that selecting the same source file could potentially create infinite shortcuts.
Added a function which checks the shortcuts in the folder, and tries to find one with the same target as the file the user is trying to set.
If found it returns that .lnk and does not create a new one.
So still possible to have multiple .lnks(for different file sources), but not possible to make multiple .lnks for the same original file.