Fixed case sensitive check for matching files (now insensitive).
When typing out a file like "E:\Background.mp4", casing did matter so if there was already a file named "background.mp4" in the directory it would not check it.

This mostly affects manually typing out files as dropping/selecting should return the correct path already. But nonetheless this is a big storage/small performance save if the check returns true, so might as well make it case insensitive to avoid possible mismatches solely due to capitalizations.