Non-named (000001 type) backgrounds now have a global counter.
Added a setting newBackgroundID which is a saved counter setting for backgrounds which are created with no name (000001, 000002, 000003 etc.).

This is mostly implemented to keep newer backgrounds made with no name as higher numbered IDs. Instead of replacing older numbered (but deleted) IDs. For example if you have 000001-000008, then delete 000002, the next created is 000009 instead of 000002.

Also added a dumb check where the number ID can go beyond 6 characters if someone is dumb enough to set it to 999999 via the settings.json file.