Added UI for Delete Background contextMenu option.
Added the Delete Background contextMenu option. Is not fully implemented, just shows a confirmation window, unselects the background if selected and logs "attempting to delete background".
Need to implement the backend to actually delete it.

Also limited it directly to 1 background deletion at a time. I was going to allow multiple deletion, but then when just testing how it worked, realized a common flaw with that. Basically I had selected a bunch of backgrounds I had wanted to delete but since I knew I was going to delete a few I started off the selection with a ctrl + click, thus I had my current background selected + all the backgrounds I wanted to delete. This is slightly dangerous because if you don't realize it you are also deleting your currently showing background.
As such at least for now just sticking with the delete 1 at a time. If you really want to delete a lot and you know the IDs you can manually select them from the /backgrounds/ folder, delete them and re-open backgroundSelect.