WIP pages and filtering.
Added a paging system and different formatting for getBackgroundSummaries. It now takes an offset ( like page number * max page entries), a limit: # of backgrounds to return, and a search "" to apply a filter to (this currently works for both name and ID.)

The paging system ensures that it only really loads what is necessary, while still managing a way to view all of the backgrounds in order, and reload pages.

Known issue: need some special case for the currently set background to stay in memory, and keep the sidebar (also needs a skip to this background button which would skip to the current page with the selected background).
Page size is currently 10 (for testing with more pages without having to create hundreds of random backgrounds).