Indexing existing backgrounds now asks for where to sort them.
Copying folders which already have a bg.json into the /backgrounds/ folder now asks the user if they want to import these as new, or import based on modified date.

Example use case of this is say that a user moves backed up or background files from 1 file to the other. The user might not want to have their format and all their files get moved around. So a small window now appears to ask the user if they want to import the backgrounds as new (appear first), or "import with last modified date".

Looking at this technically this will still slightly alter the exact format, as last modified date will trigger even when doing simple things like just changing the tags.

Looking into a better way to ensure this works as expected.