Added "indexed" local key to keep track of when it was indexed.
Looked for a way to solve the issue with the last commit of loading from last modified date (of bg.json). The problem with this is that it would not stay consistent for instance changing tags or even just going into edit Background then saving it with no changes would modify the file, but not update it's index time in Backgrounds.json. Therefore importing based on modified time creates inconsistencies.

This indexed field is the first step towards fixing this. Basically it will only ever add an indexed on saving with no indexed set.