BackgroundSelect now re-fetches pages after new indexing.
If BackgroundSelect triggers the index for new files and shows the import smallWindow confirmation, it now sends a message to re-fetch after indexing.
The problem was it would halt all indexing until the user told it what approach to take, then complete the indexing. But by then the BackgroundSelect had already fully loaded and displayed a page.
This now tells it to re-fetch the page on completing the logic in indexBackgrounds(), thus showing the updated backgrounds (if they appear on the page).