Add ipc call IdToBackgroundFolder + magnifying icon in Edit Background.
Added an ipc call to ensure that a backgroundFolder exists for (id).

Also added a magnifying glass icon in Edit Background which uses this.
Basically this is another folder icon but it defaults to the backgrounds/id/ folder.

Not incredibly helpful right now but when this is changed to open folders, this allows much faster bg.json building from an existing folder which does not yet have a bg.json.