Added both magnifyingGlasses and functionality.
Added both magnifying glasses allowing the user to easily select from in the folder to set the background path or the icon.

Still need to update this to allow opening folders themselves which are dropped onto BackgroundSelect or saved in /backgrounds/ (but no bg.json). That will probably be the next focus.