Better page swapping.
Instead of just raw text and a Prev and Next button it now displays several pages to the left/right As well as always showing 1st and last page.
Allows up to 2 pages on either side of the current selected page. If you are more than 3 from the beginning/end it will add a "..." which when clicked will let you type the number of the page to go to.