Added External Paths setting.
Added a setting External Paths. This is a name rechange from masterPaths which was previously stored in backgrounds.json but never used. I feel it better describes what the user expects.

This will be used to effectively select external "backgrounds/" like folders which hold a bunch of backgrounds but can be stored on different disks or areas. This will basically blend these backgrounds to be used just like any other backgrounds, and you will be able to save to a certain path or default etc.

However I am still conflicted on how backgrounds which already exist but do not have a bg.json are to be added. For instance when you add a folder that has the icon + bfFile but no bg.json.