Indexing extenal Paths (still extremely early in development).
indexBackgrounds now finds and loads/displays the backgrounds found in externalPaths (any folder in that directory with a bg.json).

Does not fully work, for now just displays it as a selectable background with the name = ${externalPath}\${folder_name}

Still need to implement a bunch of situational path conversion and create the reference system for externalPaths.