Fix id in BackgroundSelect and externalPaths getting overwritten.
Fixed a problem where saving the backgrounds.json via saveSettingsData would reset the externalPaths in backgrounds.json to [] if not explicitly called with data.externalPaths.

Changed the display in BackgroundSelect to remove the leading ID sequence "ext::#::" if the id resolves to that. Still need a way to actually fetch the name from the bg.json.