External Paths Backgrounds now load.
Fixed the idToBackgroundPath/idToBgJson/idToBackgroundFolder to correctly adjust external paths (paths which start with "ext::0::") to correctly resolve to the adjusted path.

This fixed the background to properly work, but loading the BackgroundSummary is still wrong. So while it will correctly display the background the BackgroundSelect is still barebones with just the file name, no preview or other information is correctly loaded.