Improving Deleting backgrounds in use.
Found an issue when deleting actually saved video backgrounds (not .lnks) which were recently in use.
Basically the problem is electron needs to unload it before it can attempt deletion. As a start to this it now previews the fallback, and shows a small window warning that it might have failed to still being in use.
This is not 100% ideal but an improvement and diagnostic for helping the user understand why the deletion fails even after giving program admin permissions.
Ideally we don't even attempt to show the confirmations for admin as this is just a shell.trashItem thing (which giving admin does not change anything as this prompt shows up assuming there is another issue which it needs admin for and not just that it cannot move it due to being in use.)