Fixed magGlass not opening correct folder in EditBackground for External Path.
Fixed the openFileDialog ipc call to accept absolute paths (while maintaining its use in local paths).
Now when called with a local (non-absolute) path it appends like so: /AppData/Roaming/AltDesktop/{filePath}

Or when called with an absolute path like so:
E:\test\000001
It now opens up the fileDialog for this absolute path.

Utilizing this the handleGlassClick for EditBackground has been updated to call with an absolute path gotten via idToBackgroundFolder(id) to properly open at the correct directory for External Paths.