Basic filter search panel.
The idea behind this is that when you click filter search it replaces the background-details-panel on the right, such that you can still see the backgrounds that appear when you apply things like tags, source directory and other saved stuff.

Still very barebones, has a few options but these are just to test toggling. Clicking an option just logs option set to "true/false".