Filter Search now loads from PUBLIC_TAGS.
Filter Search now creates a checkbox for each tag in PUBLIC_TAGS. Allowing users to turn on/off tags. These are all set to checked on opening BackgroundSelect (Might change this to store checked/unchecked statuses in the future so that when you close and re-open it the filter remains the same instead of currently which defaults back to all checked on opening it.).