Filter improvements. (still a long way to go)
Slightly improved filter.
Adding/removing tags now updates the filter.
Still a bunch to work on. Ideally the system goes like this:
Opening it up, nothing is checked.
Checking a tag/option sets it to strict filter where it only shows items which have all selected tag/options.
Unchecked are wildcards.
Double click (or some other way) to exclude a tag/option. This is strict where it removes all possible backgrounds with that tag/option.

Problems with current implementation:
All default to checked(see reason below).
When all set false, closing menu with a background that appears when all are checked removes the background.
Checked boxes do not work as exclusive, they work as a whitelist where only items with those tags (and all of the tags enabled) appear.