PartialFix: Two icons which share an ID.
Partial fix for two icons that share an ID both returning via getBackgroundSummaries.

Basically the problem is that the search for the handleSelect only searches via: id, which since the search function is quite broad, anything which matches that id case is also returned, so another background with the same id in an external path can also be returned, and it would only return the first one it found. So it was possible to request via ID and actually select another background instead of the actually visibly selected background. This would then set the background-details-panel to the backgroundSummary return (hence it would show the data of a non-selected background).

Not completely fixed but a step in the right direction.