Improved fetchPage and filterOptions updating.
Realized the mistake which made me remove the useEffect for filterOptions in an earlier commit today. It was just getting reset every time (before includeTags was moved into the fetchPage()).

Reverted, improved, and optimized the logic.
Now when adding a filter it updates includeTags, and sets the page to 0 (first page) if it is not on the first page which will then load the filtered first page. If on the first page it just applies the filter and reloads the first page.

For the user:
Adding filters will always reset the page to the first page.
Slightly optimized how the tag filters are applied.