Strict include tags. (Backgrounds must have all selected tags).
Converted from the OR logic which would return all backgrounds which had any of the tags which were in includeTags to AND logic where the background must have ALL tags in order to appear.

The OR logic was a remnant of a different attempt at design for the includeTags and excludeTags where all tags were considered included by default.

Still a long way to go, spent a lot of the time yesterday and today just going through what the process of this search should be. Also I think I am going to convert it to modifying the search operators instead of passing includeTags and excludeTags. For example a search like "Background" with the tags: "Image" "Space" would send a search variable like "Background tag:Image tag:Space"