Fixed race condition showing wrong backgrounds on opening BackgroundSelect
Fixed a race condition between the 3 useEffects, all 3 of which fired on launch. Now the only one that fires on launch is the one with no dependency, which gets the page of the saved background and displays that page (or the 1st page if no saved background).

Before based on the race condition it could display the wrong page, details under the correct page (example opening it with a background saved on page 2 could display it as being on page 2, but it's not there and instead all the backgrounds are page 1 backgrounds).