Advanced search filtering. (Still WIP).
Improved search to extract out tags in search and append them to the includeTags generated from the filter menu.

For example a user can now type:
background tag:image -tag:space

Which will find all backgrounds which:
Name, ID, or tags contain "background".
Has the "image" tag.
Does not have "space" tag.

Still no way in filter panel to exclude tags (will require a bit of work).

Originally I had entirely written this for the main process. But after some thought I decided to remake this in the UI. It's not a huge performance improvement, but in the case I ever go to add dropdown text predictions like this:
tag:im       (then a dropdown menu with these):
tag:image
tag:imagination
tag:im-a-cool-tag   etc.
It is nicer to just keep all the logic together in the UI.