Tag suggestions on typing "tag:"
When typing in the search menu "tag:" it will now suggest based on the text after tag:

For example typing in:
tag:t
will suggest:
tag:These
tag:Tags
(currently it will suggest any public tags (the tags which appear in filter search panel at the moment)).

This will be expanded on for local tags when local tags gets fully implemented. (Still a lot of work).