Only show Public tags in details panel.
Now when you have selected a background in the details panel it will only show you the tags which are in PUBLIC_TAGS. Eventually this will be expanded with local tags when you actually save local tags.

This is partially just to avoid issues where say a user shares one (without going through the export method I plan to release), and it has a bunch of user's local tags already set. In this case it will still keep the tags, in case the user ever re-adds a local tag to match it, but will only ever display the tags within the users public/local tags.

Also exported the PUBLIC_TAGS to publicTags.ts in a "shared" folder. So that it can be referenced by both the main process and the renderer.