Fixed config setup not loading new transpiled code.
When I added the "shared" folder for publicTags.ts this created an issue where the newly transpiled code.
dist-electron file was using it like this:
/utils
/windows
(all other electron js files).

But after adding shared it instead now compiled code like this:
/electron (updated transpiled code).
/shared
(Then the original code above was still in the root of the folder).

But due to the tsconfigs it was still trying to load the base dist-electron js code over the newly transpiled.

It could have been set up such that it can directly get the code from the /electron file, but the simple solution is to just remove the high src root shared folder in general and just import it in the renderer from the electron directory.

Took a few commits to notice the problem due to it only affecting the main process not the ui. whoops.