Removed /desktop directory. desktopIcons.json and settings.json now exist in root folder.
The two .json files now reside in the root AltDesktop folder.

Does not make much sense to hide them in the /desktop folder. Especially the settings.json as it only controls a few things in the desktop but has settings for the whole program.

This also more easily allows a user to understand the important .json files which correspond to their program setup.