Local tags filtering on getBackgroundSummaries.
Now when displaying localTags for a background it now only shows saved localTags (For now only exist by manually editing the settings.json).

This essentially allows other non-localTags to remain saved for a background, but only show the relative localTags for the user.
For instance if a user unproperly exported (just zipped the file instead of going through the future export process) and shared a bg.json with localTags which were not in the new users localTags they would not be removed, but would not appear when the new user selects the background/Edits tags in EditBackground.