Don't keep indexing for tags which do not exist for user.
Now when creating the backgrounds.json during indexing it only saves the tags which actually exist either in the public tags, or is a local tag.

This is to stop accidental filtering on tags which do not exist in the users scope. For example if the user had a background exported improperly from another user and it had the localTag: "cool-tag" then the user could type in "cool" or "cool-tag" or even "tag:cool-tag" and it would still appear despite not showing that it had these tags.