Search tag: is now case insensitive
Now the includeTags and excludeTags are dropped down to lowercase on call.
This makes it case insensitive. i.e. tag:Video and tag:video both return the same.
Secondly this fixed a bug where searching tag:Video was not returning despite being the autofill suggestion (it was sensitive and was looking for tag:video).

Moving forward since this is insensitive it is important when allowing the user to add tags that the tag must not exist (especially not with case differences). To avoid possible issues like user adds a local tag "image" which now triggers unintended behavior due to also having a public tag "Image"