Added "Create Background" button to empty backgroundSelect panel.
Added a button which now opens EditBackground to the empty (no selected background) in backgroundSelect.

Known issue: this is currently calling without an ID. which causes problems on saving. Gotta re-find where auto ID creation is and attach it to this to properly allow creating/saving backgrounds when opened via this.