MinHeight/Width for BackgroundSelect.
Added a minWidth and minHeight to backgroundSelect.

This is set small to accommodate for smaller screens, basically the required minimum width/height for the window to be even usable.

Will implement size scaling for icons in backgroundSelect eventually. For now it is defaulting to large icons (was originally to make sure that the preview icons appeared correctly at the right size).