Changed format of localTags. + addLocalTag ipc call
Changed the format of localTags from string[] to LocalTag[] which now stores objects of localTag which look like this:
    {
      "name": "tag",
      "category": "fun",
      "favorite": false
    }

This will be better implemented later to allow more customization and search from the localTags.

Known issue, there is still a few locations where it takes localTags and expects it to be a string[]. I am chasing these down to fix them, but as it stands it does correctly add a local tag.

Here is an example ipc call:
await window.electron.addLocalTag({
  name: "moon",
  category: "sky",
  favorite: false,
})

Also exported logic for saveSettingsData ipc call to a function in settings.ts which now does 90% of the same stuff, and can be called via other ipc calls/functions.