Fixed hard error:  TypeError: tag.toLowerCase is not a function
Fixed a reference to localTags trying to lowerCase it. It is already set to be stored as lowercase, thus does not need conversion.

Still cannot fully add/remove the localTags from individual backgrounds. They appear as options and save to the bg.json but are not properly tracked in backgrounds.json or sent as summaries for BackgroundSummaries.