Fixed backgroundSummaries not returning proper localTags.
The problem with tags displaying both in background-details-panel and EditBackground was just a problem with getBackgroundSummaries check to only return localTags which are defined within localTags setting.
Since the localTags setting format was changed from string[] to LocalTag type it just needed updating.

localTags now properly display in background-details-panel and EditBackground.