More local tags CSS adjustment.
Not 100% satisfied with this but I think this is close to as optimal as this awkward way to address it can be.
The real problem with this is that there is almost no continuity between opening EditBackground when adding/removing tags from categories.
It basically just re-arranges as it feels like it, such that if you add a tag to a category you could open it to an completely differently arranged format.
I am looking into other options but as this is the best I could make it so far with this setup will just place this here in case the other options are worse :).