Super rough new Edit Background Approach.
Few things to note:
Not 100% set on this setup.
This setup is built more toward categories being large groupings with dozens of tags each (ideal). While still allowing the user to show/hide categories.
Definitely seems more simple to edit as the user already sees this details panel when selecting backgrounds.
Allows more space for tag editing to more easily show more tags at a given time (granted the width was increased from 750 to 910 as well...

Still quite a lot to clean up, like for example down arrow on save background as, and the scroll bars for the tag section and the details section not taking into account the footer, thus the very last row of tag section is unselectable.