Fixed EditBackground scrollbars not accounting for footer.
Fixed the EditBackground window, where both the background panel and the tag panel were not accounting for the footer, thus the bottom of the scrollbar would appear below the footer.
This also for the tag panel meant that it would be impossible to click the last row of tags as it appeared behind the footer.

Also fixed some classes within the edit-background-panel.