Edit Tags window fixes (appear on top, ipc call to open)...
The Edit Tags window is now firstly a child of the activeSubWindow (how we will be opening it via the ui) Or the Main window (as a fallback just to ensure it keeps on top of the window).

Added ipc call openEditTagsWindow() which has been attached to the "Add Tags" button in Edit Background. (will be accessible in more locations later).

Added a utility to util.ts (getMainWindow()). I have always been doing it in a roundabout way where I was passing mainWindow from item to item or getting it some other way like by searching through active windows... Should have added this a long time ago.
Will probably convert some of the roundabout methods over to this soon, but haven't converted any as of now.