Modalized EditTags.
Simplicity first.
I could continue adding it as a separate window and chase down a whole bunch of bugs and fix as many as I could find. Or I could just swap to a modalized EditTags component.

Swapping to a Modalized EditTags still looks clean, vastly simpler, and if anything likely quicker/easier for a user to interact with.

And since it is a component it can be loaded in separate windows very easily. (Will likely add it to settings, if not also BackgroundSelect).