Drag/drop between categories (early stages).
Can now drag/drop tags from one category into another.
For now it only logs dragged x tag into y category.
But updated the css to show drag/droppability.
Also fixed a bug where clicking on the button in the category name would cause it to double call (thus not work).

Also fixed up some of the css like making the category names appear on the right and added a backdrop for the category headers.
(Still not 100% satisfied with the look).