syncAndGetCategories helper for settings
This will be used to sync and update the categories Setting.
It will add any new category found in localTags which is not defined in categories setting.
This ensures all categories are accounted for whenever called.

Idea for this strategy is allow categories to exist even without tags, and the ordered array will essentially control the order of the categories. For instance new categories will be added to the beginning, thus appearing first (it is assumed newer categories would be the most desired by the user). Also will implement some method to move the categories around, and change name, add or delete categories.