Fixed two infinite loops/memory leak.
Fixed a useEffect which set a useState which then was also a dependency of itself (infinite loop).

Fixed another unintentional re-calling of a function over and over (accidentally added outside a useEffect thus infinite loop).

Still looking for 1 more bug like this in EditBackground, as it is still using much more cpu than needed. ~10% vs 3-4% in any other menu.