I'm dumb (fixed mistake in previous commit infinite loop).
Forgot the dependency array. So this was running infinitely. Ironically the fix was also the thing which caused a new albeit lesser infinite loop....

Went through several other useEffects removing them to see if they were causing it, just to realize that the thing I thought I fixed was still infinite...