EditTags -> AddTag refactor.
Originally created EditTags as a separate window, but now that it is modalized it can be displayed within windows, and only really is used for adding tags, so the name AddTag better suits it.

Also cleaned up some old references like Route which added a route to /edit-tags which no longer exists as the window was removed and the component modalized.