Drag/Drop categories Rough draft (not 100% set in stone).
This uses a kind of rough approach for re-arranging the categories.
Has some major flaws with it, it's usable/workable, but kind of rough.
The category changes are only for current EditBackground (does not send/save changes).

Mostly commiting this so I have something to come back to if a modalized simplified version of this does not work.