Modalized Edit Categories prototype.
A basic modalized component which appear when clicking "Categories" button.
For now just has an input for a new category (does not actually fully create one via settings, just visually creates one for this component on clicking add category).

Then has a list of the categories in order.
In the future, this would be a much simpler version of the last commit, where you can drag categories to re-arrange them, rename them, delete them etc.