Fixed closing EditCategories not re-ordering categories.
Closing the EditCategories (Categories button) now properly reloads categories (to use the possibly updated order) when closed.

Also fixed a Add Tag closing via clicking off the modal to refresh the tags + categories as well (not 100% needed, but helps ensure the user is looking at the up to date version of the tags + categories).