Properly reloading categories on changes.
Fixed deleting a category to actually reload the saved categories to prevent the category sticking around as a local state only category.

Better ensured that the user sees the proper up to date saved status of categories positions/names etc. whenever they make changes.
( Had no problems before, but this refreshing categories after making changes ensures 100% that what is saved, is displayed, and removes the chances for a local state to say, "sure this is fine" when the saved state is different).