Favorites for tags now apply and stick.
Now when clicking on the star for a tag it updates the tag to favorite true/false.

Not 100% sure if I want to add a separate area for the favorited tags or just leave it like this where it is a visual indicator.
I originally wanted a separate space at the top for the favorites, but with the uncategorized also being there not sure if it would be overwhelming with multiple areas essentially stickied to the top. Honestly might just remove the favorite system entirely.