Fixed useEffect adding multiple "backgrounds-updated" listeners.
Fixed a useEffect which added a listener every time the user changed pages thus adding an infinitely expanding set of listeners.

Also removed fetchPage's parameter "page" as it is a local variable name for a global variable with the same name. The reasoning behind removing this is that you only call fetchPage by itself to reload the current page. Any other page instead of calling fetchPage you just change the page and the useEffect with [page, search] as a dependency will auto fetch the new page.

If you notice all the fetchPage calls before were simply just fetchPage(page) <- the global useState anyways.