Categorized public tags.
Added categories for Public tags. Which will allow for showing/hiding the tags.

Note: public tags in the filter search is pretty barebones, and hard to see what box is triggering what for the categorized public tags. But was barebones before so not much has changed, been focused on the adding/removing and the UI for that before I get around to improving the filter search.