categories now saved as Record<string, boolean>
Categories is now created as a key: value pair of name: boolean.
The saved boolean will be used for essentially for a value of being open/collapsed. So for instance when you collapse a category, the close the window and re-open the window it will essentially remember the last state of the categories, such that when you re-open it you don't have to go and minimize all the categories you made but haven't used recently.

This second part obviously is not fully implemented, and needs more work, but this is a seemingly stable version which uses the updated key: value pairs for categories.