Settings now checks types + categories migration.
Since I changed the type of the "categories" setting last commit. I realized that it was not ever performing checks on the type of the settingsData which could lead to some problems.

Changed the ensureDefaultSettings() to now replace unproperly typed settings. Also added a migrateCategoriesSetting() which will migrate the old "categories": [] to a "categories": Record<string, boolean> automatically.

So if somehow you are using this and didn't run the last commit and change any categories, you can run it now and it will auto migrate your old categories to the proper new setup.