If anything migrates, or resets to default, creates backup settings_old.json.
Added a backup file "settings_old.json" which now gets created whenever the user launches the program and ensureDefaultSettings resets a setting to default, or migrates a setting (whenever the settings is updated).

This gives the user the chance to see something different/reset, and have a chance to manually go through and migrate it/fix the problem.

Also added a disclaimer small window to notify the user this has been created.