Fix EditCategories not saving with "" category.
Fixed a case where the removal of "" as a listed category in EditCategories causes it to not be saved (removes it) from the categories setting.

Now saves an object for the "" category on creation, and appends it before calling saveSettingsData, this ensures that we can still edit the categories orders, but keep the status of the "" category expanded/hidden value.