Fixed EditBackground incorrectly saving when opened with show: false.
Fixed a case where opening the EditBackground with "show": false would correctly show the public tags section as hidden, but when toggling to show the public tags section it would re-save as "show": false. Then after that it would work as normal leaving you with a saved result which was the opposite of the actual state of the public tags section.