Added ipc "renameLocalTag" command.
Added a command to rename local tags.

call via: await window.electron.renameLocalTag("old-tag-name", "new-tag-name");

Updates the tagIndex for the new name, and updates all the bg.jsons for changing the old tag name to the new tag name.

Now just need to attach this into the rename contextMenu for right clicking on a local tag.