RenameTagModal.
Added a modal based on the AddTag modal which appears when you click the Rename in the contextMenu of a local tag.

It allows renaming of the tag via the previously added ipc command.

Still somewhat barebones, but seems to be working for the most part.

KNOWN ISSUE: when renaming a tag which is applied, it then does not apply it to the new tag name (after reloading local tags). Basically if you have a tag applied un-apply it, then rename it, then re-apply it. Or else it will unapply it on save.