Fixed renameTagModal not re-selecting updated tag name.
Fixed a case where renaming an already selected tag would no longer show it as a local Tag and would not select the tag with the new name.

This caused a few problems.
1. The old tag name still existed in the Background Summary, thus saving would still keep the old tag name as a selected tag, even though that tag no longer exists.
2. Saving after renaming without re-selecting the new tag name would result in losing the tag for that background.

Simple solution to this is to just update the Summaries localTags to match the new tag name. This ensures that we remove any old names the Summary which will be saved, and ensures that the loadLocalTags properly displays the new tag as selected and saves properly.